This commit was manufactured by cvs2git to create tag 'SL03f'.

+++ StRoot/StSpinMaker/StppJetAnalyzer.cxx
@@ -202,7 +202,7 @@ bool StppJetAnalyzer::acceptJet(StProtoJet &pj)
 	    (pj.pt() > mJetPtMin)
 	    && (fabs(pj.eta()) < mJetEtaMax)
 	    && (fabs(pj.eta()) > mJetEtaMin)
-            && ((int)pj.numberOfParticles() >= mJetNmin)
+            && (pj.numberOfParticles() >= mJetNmin)
 	    );
 }
